third_party/WebKit/LayoutTests/http/tests/devtools/editor/text-editor-char-to-coordinates-expected.txt

CONSOLE WARNING: line 3078: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Test editor cursorPositionToCoordinates and coordinatesToCursorPosition API


Running: testCoordinatesToCursorPositionOuter
Request char at (-1000, -1000)
Result:null

Running: testTextToCoordinatesCornerCases
Request (-1, 0):null
Request (100, 0):null
Request (0, -1):null
Request (0, 100):null
Request (-100, -1100):null

Running: testInverseRelationTextToCoordinates
Testing line 'function foo(a, b) {'
Testing line '    var f = /*.[a]/.test(a);'
Testing line '    return f;'
